requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pyotp==2.9.0
qrcode==7.4.2
pillow==10.3.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
aiosqlite==0.20.0
greenlet==3.0.3
websockets==12.0
python-dateutil==2.9.0.post0
